chore: axios interceptors 정리

+++ src/api/axios.ts
@@ -1,8 +1,6 @@
 import axios from 'axios';
-import { redirect } from 'react-router-dom';
 import { environments } from '~/config/environments';
 import CONSTANTS from '~/constants';
-import { ERROR_MESSAGES } from '~/constants/errorMessage';
 import { deleteCookie, getCookie, setCookie } from '~/utils/cookie';
 
 export const resumeMeAxios = axios.create({
@@ -30,46 +28,25 @@ resumeMeAxios.interceptors.response.use(
     const statusCode = error.response.status;
     const { code } = error.response.data;
 
-    switch (statusCode) {
-      //FIXME: 에러 코드 획정되면 code 수정하기
-      case 400:
-        if (code in ERROR_MESSAGES) {
-          if (code === 'INVALID_ACCESS_TOKEN') {
-            deleteCookie(CONSTANTS.ACCESS_TOKEN_HEADER);
+    if (statusCode === 400 && code === 'INVALID_ACCESS_TOKEN') {
+      deleteCookie(CONSTANTS.ACCESS_TOKEN_HEADER);
 
-            const originalRequest = error.config;
+      const originalRequest = error.config;
 
-            const refreshToken = getCookie(CONSTANTS.REFRESH_TOKEN_HEADER);
+      const refreshToken = getCookie(CONSTANTS.REFRESH_TOKEN_HEADER);
 
-            if (!refreshToken) return;
+      if (!refreshToken) return;
 
-            originalRequest._retry = true;
-            originalRequest.headers[CONSTANTS.REFRESH_TOKEN_HEADER] = refreshToken;
+      originalRequest._retry = true;
+      originalRequest.headers[CONSTANTS.REFRESH_TOKEN_HEADER] = refreshToken;
 
-            const { headers } = await resumeMeAxios(originalRequest);
+      const { headers } = await resumeMeAxios(originalRequest);
 
-            const newAccessToken = headers[CONSTANTS.ACCESS_TOKEN_HEADER];
+      const newAccessToken = headers[CONSTANTS.ACCESS_TOKEN_HEADER];
 
-            setCookie(CONSTANTS.ACCESS_TOKEN_HEADER, newAccessToken);
-          }
-        }
-        break;
-      case 403:
-        //에러 코드 추가되면 내용 추가하기
-        if (code === '멘토가 접근할 수 없는 내용입니다.') {
-          redirect('/');
-        }
-        break;
-      case 404:
-        if (code === '페이지 없음 코드') {
-          redirect('/'); //FIXME: 404페이지로 리다이렉트
-        }
-        break;
-      default:
-        if (statusCode >= 500) {
-          console.error(error.response);
-        }
+      setCookie(CONSTANTS.ACCESS_TOKEN_HEADER, newAccessToken);
     }
+
     return Promise.reject(error);
   },
 );
